+++ Changelog.md
@@ -1,3 +1,9 @@
+# Master (unreleased)
+
+* Add mutation from `Hash#[]` to `Hash#key?` (#511)
+* Add mutation from `Hash#fetch` to `Hash#key?` (#511)
+* Add mutation from `#at` to `#key?` (#511)
+
 # v0.8.8 2015-xx-xx
 
 * Drop support for rspec-3.2